refactor yurtadm join
