+++ pkg/util/kubernetes/kubeadm/app/constants/constants_unix.go
@@ -1,3 +1,4 @@
+//go:build !windows
 // +build !windows
 
 /*
