+++ pkg/util/kubernetes/kubeadm/app/preflight/checks.go
@@ -25,7 +25,6 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"path/filepath"
 	"runtime"
 	"strings"
 
@@ -40,17 +39,9 @@ import (
 	utilsexec "k8s.io/utils/exec"
 	utilsnet "k8s.io/utils/net"
 
-	kubeadmapi "github.com/openyurtio/openyurt/pkg/util/kubernetes/kubeadm/app/apis/kubeadm"
 	kubeadmconstants "github.com/openyurtio/openyurt/pkg/util/kubernetes/kubeadm/app/constants"
 	"github.com/openyurtio/openyurt/pkg/util/kubernetes/kubeadm/app/util/initsystem"
 	utilruntime "github.com/openyurtio/openyurt/pkg/util/kubernetes/kubeadm/app/util/runtime"
-)
-
-const (
-	bridgenf              = "/proc/sys/net/bridge/bridge-nf-call-iptables"
-	bridgenf6             = "/proc/sys/net/bridge/bridge-nf-call-ip6tables"
-	ipv4Forward           = "/proc/sys/net/ipv4/ip_forward"
-	ipv6DefaultForwarding = "/proc/sys/net/ipv6/conf/default/forwarding"
 )
 
 // Error defines struct for communicating error messages generated by preflight checks
@@ -715,91 +706,6 @@ func (ncc NumCPUCheck) Check() (warnings, errorList []error) {
 		errorList = append(errorList, errors.Errorf("the number of available CPUs %d is less than the required %d", numCPU, ncc.NumCPU))
 	}
 	return warnings, errorList
-}
-
-// RunJoinNodeChecks executes all individual, applicable to node checks.
-func RunJoinNodeChecks(execer utilsexec.Interface, cfg *kubeadmapi.JoinConfiguration, ignorePreflightErrors sets.String) error {
-	// First, check if we're root separately from the other preflight checks and fail fast
-	if err := RunRootCheckOnly(ignorePreflightErrors); err != nil {
-		return err
-	}
-
-	checks := []Checker{
-		DirAvailableCheck{Path: filepath.Join(kubeadmconstants.KubernetesDir, kubeadmconstants.ManifestsSubDirName)},
-		FileAvailableCheck{Path: filepath.Join(kubeadmconstants.KubernetesDir, kubeadmconstants.KubeletKubeConfigFileName)},
-		FileAvailableCheck{Path: filepath.Join(kubeadmconstants.KubernetesDir, kubeadmconstants.KubeletBootstrapKubeConfigFileName)},
-	}
-	checks = addCommonChecks(execer, "", &cfg.NodeRegistration, checks)
-	if cfg.ControlPlane == nil {
-		checks = append(checks, FileAvailableCheck{Path: cfg.CACertPath})
-	}
-
-	addIPv6Checks := false
-	if cfg.Discovery.BootstrapToken != nil {
-		ipstr, _, err := net.SplitHostPort(cfg.Discovery.BootstrapToken.APIServerEndpoint)
-		if err == nil {
-			checks = append(checks,
-				HTTPProxyCheck{Proto: "https", Host: ipstr},
-			)
-			if ip := net.ParseIP(ipstr); ip != nil {
-				if utilsnet.IsIPv6(ip) {
-					addIPv6Checks = true
-				}
-			}
-		}
-	}
-	if addIPv6Checks {
-		checks = append(checks,
-			FileContentCheck{Path: bridgenf6, Content: []byte{'1'}},
-			FileContentCheck{Path: ipv6DefaultForwarding, Content: []byte{'1'}},
-		)
-	}
-
-	return RunChecks(checks, os.Stderr, ignorePreflightErrors)
-}
-
-// addCommonChecks is a helper function to duplicate checks that are common between both the
-// kubeadm init and join commands
-func addCommonChecks(execer utilsexec.Interface, k8sVersion string, nodeReg *kubeadmapi.NodeRegistrationOptions, checks []Checker) []Checker {
-	containerRuntime, err := utilruntime.NewContainerRuntime(execer, nodeReg.CRISocket)
-	isDocker := false
-	if err != nil {
-		fmt.Printf("[preflight] WARNING: Couldn't create the interface used for talking to the container runtime: %v\n", err)
-	} else {
-		checks = append(checks, ContainerRuntimeCheck{runtime: containerRuntime})
-		if containerRuntime.IsDocker() {
-			isDocker = true
-			checks = append(checks, ServiceCheck{Service: "docker", CheckIfActive: true})
-		}
-	}
-
-	// non-windows checks
-	if runtime.GOOS == "linux" {
-		if !isDocker {
-			checks = append(checks, InPathCheck{executable: "crictl", mandatory: true, exec: execer})
-		}
-		checks = append(checks,
-			FileContentCheck{Path: bridgenf, Content: []byte{'1'}},
-			FileContentCheck{Path: ipv4Forward, Content: []byte{'1'}},
-			SwapCheck{},
-			InPathCheck{executable: "conntrack", mandatory: true, exec: execer},
-			InPathCheck{executable: "ip", mandatory: true, exec: execer},
-			InPathCheck{executable: "iptables", mandatory: true, exec: execer},
-			InPathCheck{executable: "mount", mandatory: true, exec: execer},
-			InPathCheck{executable: "nsenter", mandatory: true, exec: execer},
-			InPathCheck{executable: "ebtables", mandatory: false, exec: execer},
-			InPathCheck{executable: "ethtool", mandatory: false, exec: execer},
-			InPathCheck{executable: "socat", mandatory: false, exec: execer},
-			InPathCheck{executable: "tc", mandatory: false, exec: execer},
-			InPathCheck{executable: "touch", mandatory: false, exec: execer})
-	}
-	checks = append(checks,
-		SystemVerificationCheck{IsDocker: isDocker},
-		HostnameCheck{nodeName: nodeReg.Name},
-		KubeletVersionCheck{KubernetesVersion: k8sVersion, exec: execer},
-		ServiceCheck{Service: "kubelet", CheckIfActive: false},
-		PortOpenCheck{port: kubeadmconstants.KubeletPort})
-	return checks
 }
 
 // RunRootCheckOnly initializes checks slice of structs and call RunChecks
