+++ pkg/util/kubernetes/kubeadm/app/util/runtime/runtime_unix.go
@@ -1,3 +1,4 @@
+//go:build !windows
 // +build !windows
 
 /*
