+++ pkg/yurtadm/util/token/token.go
@@ -1,6 +1,6 @@
 /*
-Copyright 2016 The Kubernetes Authors.
-Copyright 2021 The OpenYurt Authors.
+Copyright 2022 The OpenYurt Authors.
+Copyright 2017 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -20,57 +20,65 @@ package token
 import (
 	"bytes"
 	"context"
+	"errors"
 	"fmt"
 	"strings"
 	"time"
 
-	"github.com/pkg/errors"
+	pkgerrors "github.com/pkg/errors"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
-	clientset "k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
 	certutil "k8s.io/client-go/util/cert"
 	bootstrapapi "k8s.io/cluster-bootstrap/token/api"
 	bootstrap "k8s.io/cluster-bootstrap/token/jws"
+	bootstraputil "k8s.io/cluster-bootstrap/token/util"
 	"k8s.io/klog/v2"
 
-	bootstraptokenv1 "github.com/openyurtio/openyurt/pkg/util/kubernetes/kubeadm/app/apis/bootstraptoken/v1"
-	"github.com/openyurtio/openyurt/pkg/util/kubernetes/kubeadm/app/constants"
-	kubeconfigutil "github.com/openyurtio/openyurt/pkg/util/kubernetes/kubeadm/app/util/kubeconfig"
-	"github.com/openyurtio/openyurt/pkg/util/kubernetes/kubeadm/app/util/pubkeypin"
 	"github.com/openyurtio/openyurt/pkg/yurtadm/cmd/join/joindata"
+	kubeconfigutil "github.com/openyurtio/openyurt/pkg/yurtadm/util/kubeconfig"
+	"github.com/openyurtio/openyurt/pkg/yurtadm/util/pubkeypin"
 )
 
-// BootstrapUser defines bootstrap user name
-const BootstrapUser = "token-bootstrap-client"
+const (
+	// BootstrapUser defines bootstrap user name
+	BootstrapUser = "token-bootstrap-client"
+	// DiscoveryRetryInterval specifies how long kubeadm should wait before retrying to connect to the control-plane when doing discovery
+	DiscoveryRetryInterval = 5 * time.Second
+	// PatchNodeTimeout specifies how long kubeadm should wait for applying the label and taint on the control-plane before timing out
+	PatchNodeTimeout = 2 * time.Minute
+)
 
-// TokenUser defines token user
-const TokenUser = "tls-bootstrap-token-user"
-
-// RetrieveBootstrapConfig get clientcmdapi config by bootstrap token
-func RetrieveBootstrapConfig(data joindata.YurtJoinData) (*clientcmdapi.Config, error) {
-	cfg, err := retrieveValidatedConfigInfo(nil, data)
-	if err != nil {
-		return nil, err
-	}
-
-	clusterinfo := kubeconfigutil.GetClusterFromKubeConfig(cfg)
-	return kubeconfigutil.CreateWithToken(
-		// If there are multiple master IP addresses, take the first one here
-		fmt.Sprintf("https://%s", strings.Split(data.ServerAddr(), ",")[0]),
-		"kubernetes",
-		TokenUser,
-		clusterinfo.CertificateAuthorityData,
-		data.JoinToken(),
-	), nil
+// BootstrapTokenString is a token of the format abcdef.abcdef0123456789 that is used
+// for both validation of the practically of the API server from a joining node's point
+// of view and as an authentication method for the node in the bootstrap phase of
+// "kubeadm join". This token is and should be short-lived
+type BootstrapTokenString struct {
+	ID     string `json:"-"`
+	Secret string `json:"-" datapolicy:"token"`
 }
 
-// retrieveValidatedConfigInfo is a private implementation of RetrieveValidatedConfigInfo.
+// newBootstrapTokenString converts the given Bootstrap Token as a string
+// to the BootstrapTokenString object used for serialization/deserialization
+// and internal usage. It also automatically validates that the given token
+// is of the right format
+func newBootstrapTokenString(token string) (*BootstrapTokenString, error) {
+	substrs := bootstraputil.BootstrapTokenRegexp.FindStringSubmatch(token)
+	// TODO: Add a constant for the 3 value here, and explain better why it's needed (other than because how the regexp parsin works)
+	if len(substrs) != 3 {
+		return nil, pkgerrors.Errorf("the bootstrap token %q was not of the form %q", token, bootstrapapi.BootstrapTokenPattern)
+	}
+
+	return &BootstrapTokenString{ID: substrs[1], Secret: substrs[2]}, nil
+}
+
+// RetrieveValidatedConfigInfo is a private implementation of RetrieveValidatedConfigInfo.
 // It accepts an optional clientset that can be used for testing purposes.
-func retrieveValidatedConfigInfo(client clientset.Interface, data joindata.YurtJoinData) (*clientcmdapi.Config, error) {
-	token, err := bootstraptokenv1.NewBootstrapTokenString(data.JoinToken())
+func RetrieveValidatedConfigInfo(client kubernetes.Interface, data joindata.YurtJoinData) (*clientcmdapi.Config, error) {
+	token, err := newBootstrapTokenString(data.JoinToken())
 	if err != nil {
 		return nil, err
 	}
@@ -87,7 +95,7 @@ func retrieveValidatedConfigInfo(client clientset.Interface, data joindata.YurtJ
 	clusterName := insecureBootstrapConfig.Contexts[insecureBootstrapConfig.CurrentContext].Cluster
 
 	klog.V(1).Infof("[discovery] Created cluster-info discovery client, requesting info from %q", endpoint)
-	insecureClusterInfo, err := getClusterInfo(client, insecureBootstrapConfig, token, constants.DiscoveryRetryInterval, constants.PatchNodeTimeout)
+	insecureClusterInfo, err := getClusterInfo(client, insecureBootstrapConfig, token, DiscoveryRetryInterval, PatchNodeTimeout)
 	if err != nil {
 		return nil, err
 	}
@@ -101,12 +109,12 @@ func retrieveValidatedConfigInfo(client clientset.Interface, data joindata.YurtJ
 	// Load the insecure config
 	insecureConfig, err := clientcmd.Load(insecureKubeconfigBytes)
 	if err != nil {
-		return nil, errors.Wrapf(err, "couldn't parse the kubeconfig file in the %s ConfigMap", bootstrapapi.ConfigMapClusterInfo)
+		return nil, pkgerrors.Wrapf(err, "couldn't parse the kubeconfig file in the %s ConfigMap", bootstrapapi.ConfigMapClusterInfo)
 	}
 
 	// The ConfigMap should contain a single cluster
 	if len(insecureConfig.Clusters) != 1 {
-		return nil, errors.Errorf("expected the kubeconfig file in the %s ConfigMap to have a single cluster, but it had %d", bootstrapapi.ConfigMapClusterInfo, len(insecureConfig.Clusters))
+		return nil, pkgerrors.Errorf("expected the kubeconfig file in the %s ConfigMap to have a single cluster, but it had %d", bootstrapapi.ConfigMapClusterInfo, len(insecureConfig.Clusters))
 	}
 
 	// If no TLS root CA pinning was specified, we're done
@@ -125,7 +133,7 @@ func retrieveValidatedConfigInfo(client clientset.Interface, data joindata.YurtJ
 	secureBootstrapConfig := buildSecureBootstrapKubeConfig(endpoint, clusterCABytes, clusterName)
 
 	klog.V(1).Infof("[discovery] Requesting info from %q again to validate TLS against the pinned public key", endpoint)
-	secureClusterInfo, err := getClusterInfo(client, secureBootstrapConfig, token, constants.DiscoveryRetryInterval, constants.PatchNodeTimeout)
+	secureClusterInfo, err := getClusterInfo(client, secureBootstrapConfig, token, DiscoveryRetryInterval, PatchNodeTimeout)
 	if err != nil {
 		return nil, err
 	}
@@ -133,12 +141,12 @@ func retrieveValidatedConfigInfo(client clientset.Interface, data joindata.YurtJ
 	// Pull the kubeconfig from the securely-obtained ConfigMap and validate that it's the same as what we found the first time
 	secureKubeconfigBytes := []byte(secureClusterInfo.Data[bootstrapapi.KubeConfigKey])
 	if !bytes.Equal(secureKubeconfigBytes, insecureKubeconfigBytes) {
-		return nil, errors.Errorf("the second kubeconfig from the %s ConfigMap (using validated TLS) was different from the first", bootstrapapi.ConfigMapClusterInfo)
+		return nil, pkgerrors.Errorf("the second kubeconfig from the %s ConfigMap (using validated TLS) was different from the first", bootstrapapi.ConfigMapClusterInfo)
 	}
 
 	secureKubeconfig, err := clientcmd.Load(secureKubeconfigBytes)
 	if err != nil {
-		return nil, errors.Wrapf(err, "couldn't parse the kubeconfig file in the %s ConfigMap", bootstrapapi.ConfigMapClusterInfo)
+		return nil, pkgerrors.Wrapf(err, "couldn't parse the kubeconfig file in the %s ConfigMap", bootstrapapi.ConfigMapClusterInfo)
 	}
 
 	klog.V(1).Infof("[discovery] Cluster info signature and contents are valid and TLS certificate validates against pinned roots, will use API Server %q", endpoint)
@@ -162,16 +170,16 @@ func buildSecureBootstrapKubeConfig(endpoint string, caCert []byte, clustername 
 }
 
 // validateClusterInfoToken validates that the JWS token present in the cluster info ConfigMap is valid
-func validateClusterInfoToken(insecureClusterInfo *v1.ConfigMap, token *bootstraptokenv1.BootstrapTokenString) ([]byte, error) {
+func validateClusterInfoToken(insecureClusterInfo *v1.ConfigMap, token *BootstrapTokenString) ([]byte, error) {
 	insecureKubeconfigString, ok := insecureClusterInfo.Data[bootstrapapi.KubeConfigKey]
 	if !ok || len(insecureKubeconfigString) == 0 {
-		return nil, errors.Errorf("there is no %s key in the %s ConfigMap. This API Server isn't set up for token bootstrapping, can't connect",
+		return nil, pkgerrors.Errorf("there is no %s key in the %s ConfigMap. This API Server isn't set up for token bootstrapping, can't connect",
 			bootstrapapi.KubeConfigKey, bootstrapapi.ConfigMapClusterInfo)
 	}
 
 	detachedJWSToken, ok := insecureClusterInfo.Data[bootstrapapi.JWSSignatureKeyPrefix+token.ID]
 	if !ok || len(detachedJWSToken) == 0 {
-		return nil, errors.Errorf("token id %q is invalid for this cluster or it has expired. Use \"kubeadm token create\" on the control-plane node to create a new valid token", token.ID)
+		return nil, pkgerrors.Errorf("token id %q is invalid for this cluster or it has expired. Use \"kubeadm token create\" on the control-plane node to create a new valid token", token.ID)
 	}
 
 	if !bootstrap.DetachedTokenIsValid(detachedJWSToken, insecureKubeconfigString, token.ID, token.Secret) {
@@ -190,13 +198,13 @@ func validateClusterCA(insecureConfig *clientcmdapi.Config, pubKeyPins *pubkeypi
 
 	clusterCAs, err := certutil.ParseCertsPEM(clusterCABytes)
 	if err != nil {
-		return nil, errors.Wrapf(err, "failed to parse cluster CA from the %s ConfigMap", bootstrapapi.ConfigMapClusterInfo)
+		return nil, pkgerrors.Wrapf(err, "failed to parse cluster CA from the %s ConfigMap", bootstrapapi.ConfigMapClusterInfo)
 	}
 
 	// Validate the cluster CA public key against the pinned set
 	err = pubKeyPins.CheckAny(clusterCAs)
 	if err != nil {
-		return nil, errors.Wrapf(err, "cluster CA found in %s ConfigMap is invalid", bootstrapapi.ConfigMapClusterInfo)
+		return nil, pkgerrors.Wrapf(err, "cluster CA found in %s ConfigMap is invalid", bootstrapapi.ConfigMapClusterInfo)
 	}
 
 	return clusterCABytes, nil
@@ -205,7 +213,7 @@ func validateClusterCA(insecureConfig *clientcmdapi.Config, pubKeyPins *pubkeypi
 // getClusterInfo creates a client from the given kubeconfig if the given client is nil,
 // and requests the cluster info ConfigMap using PollImmediate.
 // If a client is provided it will be used instead.
-func getClusterInfo(client clientset.Interface, kubeconfig *clientcmdapi.Config, token *bootstraptokenv1.BootstrapTokenString, interval, duration time.Duration) (*v1.ConfigMap, error) {
+func getClusterInfo(client kubernetes.Interface, kubeconfig *clientcmdapi.Config, token *BootstrapTokenString, interval, duration time.Duration) (*v1.ConfigMap, error) {
 	var cm *v1.ConfigMap
 	var err error
 
@@ -230,7 +238,7 @@ func getClusterInfo(client clientset.Interface, kubeconfig *clientcmdapi.Config,
 		// Make sure we retry util then.
 		if _, ok := cm.Data[bootstrapapi.JWSSignatureKeyPrefix+token.ID]; !ok {
 			klog.V(1).Infof("[discovery] The cluster-info ConfigMap does not yet contain a JWS signature for token ID %q, will try again", token.ID)
-			err = errors.Errorf("could not find a JWS signature in the cluster-info ConfigMap for token ID %q", token.ID)
+			err = pkgerrors.Errorf("could not find a JWS signature in the cluster-info ConfigMap for token ID %q", token.ID)
 			return
 		}
 		// Cancel the context on success
